Move form manager edit types into the component's directory.
